+++ src/context.ts
@@ -1,5 +1,9 @@
 import type { ExtensionContext } from 'vscode'
 
+export enum WspStatKey {
+  hexColorEnabled,
+}
+
 let context: ExtensionContext
 
 export function getExtContext(): ExtensionContext {