+++ src/util/index.ts
@@ -1,5 +1,4 @@
-import { execFile } from 'child_process'
-import { EOL } from 'os'
+import { execFile, execFileSync } from 'child_process'
 import util from 'util'
 import type { CancellationToken } from 'vscode'
 
@@ -21,15 +20,10 @@ export function* mergeIterables<T>(iterables: Iterable<Iterable<T>>) {
   }
 }
 
-export async function findExeInPath(name: string) {
-  return await new Promise<string>((resolve, reject) => {
-    const finder = isWin32 ? 'C:/Windows/System32/where.exe' : '/usr/bin/which'
-    if (isWin32) {
-      name += '.exe'
-    }
-    execFile(finder, [name], { encoding: 'utf8' }, (err, stdout, stderr) => {
-      if (err) reject(err)
-      else resolve(stdout.split(EOL)[0])
-    })
-  })
+export function findExeInPathSync(name: string) {
+  const finder = isWin32 ? 'C:/Windows/System32/where.exe' : '/usr/bin/which'
+  if (isWin32) {
+    name += '.exe'
+  }
+  return execFileSync(finder, [name], { encoding: 'utf8' }).trimEnd()
 }